doc: Zip download of hello world (quickstart) sample
* moving away sample downloads from developer.lightbend.com
* include zip downloads of samples in Paradox
* this is the same sample as we have in "First example" so
  the doc content of the quickstart is redundant
* updated links from getting started
* test samples from ci
