+++ .github/workflows/publish-docs.yml
@@ -36,6 +36,7 @@ jobs:
 
       - name: Publish
         run: |-
+          scripts/prepare-downloads.sh
           eval "$(ssh-agent -s)"
           echo $AKKA_RSYNC_GUSTAV | base64 -d > .github/id_rsa
           chmod 600 .github/id_rsa
