+++ .gitignore
@@ -102,3 +102,6 @@ native/
 .tmp
 
 /fortify-scans/
+
+# attachments created by scripts/prepare-downloads.sh
+akka-docs/src/main/paradox/attachments
