+++ akka-actor-typed-tests/src/test/java/jdocs/akka/typed/IntroTest.java
@@ -26,8 +26,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public interface IntroTest {
-
-  // #hello-world-actor
   public class HelloWorld extends AbstractBehavior<HelloWorld.Greet> {
 
     public static final class Greet {
@@ -69,84 +67,9 @@ public interface IntroTest {
       return this;
     }
   }
-  // #hello-world-actor
-
-  // #hello-world-bot
-  public class HelloWorldBot extends AbstractBehavior<HelloWorld.Greeted> {
-
-    public static Behavior<HelloWorld.Greeted> create(int max) {
-      return Behaviors.setup(context -> new HelloWorldBot(context, max));
-    }
-
-    private final int max;
-    private int greetingCounter;
-
-    private HelloWorldBot(ActorContext<HelloWorld.Greeted> context, int max) {
-      super(context);
-      this.max = max;
-    }
-
-    @Override
-    public Receive<HelloWorld.Greeted> createReceive() {
-      return newReceiveBuilder().onMessage(HelloWorld.Greeted.class, this::onGreeted).build();
-    }
-
-    private Behavior<HelloWorld.Greeted> onGreeted(HelloWorld.Greeted message) {
-      greetingCounter++;
-      getContext().getLog().info("Greeting {} for {}", greetingCounter, message.whom);
-      if (greetingCounter == max) {
-        return Behaviors.stopped();
-      } else {
-        message.from.tell(new HelloWorld.Greet(message.whom, getContext().getSelf()));
-        return this;
-      }
-    }
-  }
-  // #hello-world-bot
-
-  // #hello-world-main
-  // #hello-world-main-setup
-  public class HelloWorldMain extends AbstractBehavior<HelloWorldMain.SayHello> {
-    // #hello-world-main-setup
-
-    public static class SayHello {
-      public final String name;
-
-      public SayHello(String name) {
-        this.name = name;
-      }
-    }
-
-    // #hello-world-main-setup
-    public static Behavior<SayHello> create() {
-      return Behaviors.setup(HelloWorldMain::new);
-    }
-
-    private final ActorRef<HelloWorld.Greet> greeter;
-
-    private HelloWorldMain(ActorContext<SayHello> context) {
-      super(context);
-      greeter = context.spawn(HelloWorld.create(), "greeter");
-    }
-    // #hello-world-main-setup
-
-    @Override
-    public Receive<SayHello> createReceive() {
-      return newReceiveBuilder().onMessage(SayHello.class, this::onStart).build();
-    }
-
-    private Behavior<SayHello> onStart(SayHello command) {
-      ActorRef<HelloWorld.Greeted> replyTo =
-          getContext().spawn(HelloWorldBot.create(3), command.name);
-      greeter.tell(new HelloWorld.Greet(command.name, replyTo));
-      return this;
-    }
-    // #hello-world-main-setup
-  }
-  // #hello-world-main-setup
-  // #hello-world-main
 
   interface CustomDispatchersExample {
+
     // #hello-world-main-with-dispatchers
     public class HelloWorldMain extends AbstractBehavior<HelloWorldMain.SayHello> {
 
@@ -184,19 +107,6 @@ public interface IntroTest {
       // #hello-world-main-with-dispatchers
     }
     // #hello-world-main-with-dispatchers
-  }
-
-  public static void main(String[] args) throws Exception {
-    // #hello-world
-    final ActorSystem<HelloWorldMain.SayHello> system =
-        ActorSystem.create(HelloWorldMain.create(), "hello");
-
-    system.tell(new HelloWorldMain.SayHello("World"));
-    system.tell(new HelloWorldMain.SayHello("Akka"));
-    // #hello-world
-
-    Thread.sleep(3000);
-    system.terminate();
   }
 
   // #chatroom-behavior
