+++ samples/akka-quickstart-java/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Tue Apr 16 15:24:44 CEST 2024
+gradle.version=7.6
