+++ samples/akka-quickstart-java/README.md
@@ -1,0 +1,34 @@
+# Akka Quickstart with Java
+
+## Quickstart guide
+
+This example is described in https://doc.akka.io/docs/akka/current/typed/actors.html#first-example
+
+The Hello World example illustrates Akka basics. Within 10 minutes, you should be able to download and run the example and read the Introduction to Actors documentation to understand how the example is constructed. This will get your feet wet, and hopefully inspire you to dive deeper into the wonderful sea of Akka!
+
+## How to run
+
+This template also provides build descriptors for maven and gradle. You can use any of the following commands to run the application:
+```
+mvn compile exec:exec
+```
+or
+```
+./gradlew run
+```
+
+## Template license
+
+Written in 2017 by Lightbend, Inc.
+
+To the extent possible under law, the author(s) have dedicated all copyright and related
+and neighboring rights to this template to the public domain worldwide.
+This template is distributed without any warranty. See <http://creativecommons.org/publicdomain/zero/1.0/>.
+
+The Akka family of projects is managed by teams at [Lightbend](https://lightbend.com) with help from the community.
+
+---
+
+## License
+
+Akka is licensed under the Business Source License 1.1, please see the [Akka License FAQ](https://www.lightbend.com/akka/license-faq).
