+++ samples/akka-quickstart-java/build.gradle
@@ -1,0 +1,25 @@
+apply plugin: 'java'
+apply plugin: 'idea'
+apply plugin: 'application'
+
+
+repositories {
+    mavenCentral()
+    mavenLocal()
+    maven {
+        url "https://repo.akka.io/maven"
+    }
+}
+
+dependencies {
+  implementation 'com.typesafe.akka:akka-actor-typed_2.13:2.9.2'
+  implementation 'ch.qos.logback:logback-classic:1.2.3'
+  testImplementation 'com.typesafe.akka:akka-actor-testkit-typed_2.13:2.9.2'
+  testImplementation 'junit:junit:4.13.1'
+}
+
+mainClassName = "com.example.HelloWorldMain"
+
+run {
+  standardInput = System.in
+}
