+++ samples/akka-quickstart-scala/build.sbt
@@ -1,0 +1,22 @@
+name := "akka-quickstart-scala"
+
+version := "1.0"
+
+scalaVersion := "2.13.13"
+
+resolvers += "Akka library repository".at("https://repo.akka.io/maven")
+
+lazy val akkaVersion = sys.props.getOrElse("akka.version", "2.9.2")
+
+// Run in a separate JVM, to make sure sbt waits until all threads have
+// finished before returning.
+// If you want to keep the application running while executing other
+// sbt tasks, consider https://github.com/spray/sbt-revolver/
+fork := true
+
+libraryDependencies ++= Seq(
+  "com.typesafe.akka" %% "akka-actor-typed" % akkaVersion,
+  "ch.qos.logback" % "logback-classic" % "1.2.13",
+  "com.typesafe.akka" %% "akka-actor-testkit-typed" % akkaVersion % Test,
+  "org.scalatest" %% "scalatest" % "3.2.15" % Test
+)
