+++ samples/akka-quickstart-scala/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("io.spray" % "sbt-revolver" % "0.10.0")
