+++ samples/akka-quickstart-scala/src/main/scala/com/example/HelloWorld.scala
@@ -1,0 +1,75 @@
+package com.example
+
+//#imports
+import akka.actor.typed.ActorRef
+import akka.actor.typed.ActorSystem
+import akka.actor.typed.Behavior
+import akka.actor.typed.scaladsl.Behaviors
+//#imports
+
+//#hello-world-actor
+object HelloWorld {
+  final case class Greet(whom: String, replyTo: ActorRef[Greeted])
+  final case class Greeted(whom: String, from: ActorRef[Greet])
+
+  def apply(): Behavior[Greet] = Behaviors.receive { (context, message) =>
+    context.log.info("Hello {}!", message.whom)
+    message.replyTo ! Greeted(message.whom, context.self)
+    Behaviors.same
+  }
+}
+//#hello-world-actor
+
+//#hello-world-bot
+object HelloWorldBot {
+
+  def apply(max: Int): Behavior[HelloWorld.Greeted] = {
+    bot(0, max)
+  }
+
+  private def bot(greetingCounter: Int, max: Int): Behavior[HelloWorld.Greeted] =
+    Behaviors.receive { (context, message) =>
+      val n = greetingCounter + 1
+      context.log.info("Greeting {} for {}", n, message.whom)
+      if (n == max) {
+        Behaviors.stopped
+      } else {
+        message.from ! HelloWorld.Greet(message.whom, context.self)
+        bot(n, max)
+      }
+    }
+}
+//#hello-world-bot
+
+//#hello-world-main
+object HelloWorldMain {
+
+  final case class SayHello(name: String)
+
+  def apply(): Behavior[SayHello] =
+    Behaviors.setup { context =>
+      val greeter = context.spawn(HelloWorld(), "greeter")
+
+      Behaviors.receiveMessage { message =>
+        val replyTo = context.spawn(HelloWorldBot(max = 3), message.name)
+        greeter ! HelloWorld.Greet(message.name, replyTo)
+        Behaviors.same
+      }
+    }
+
+  //#hello-world-main
+  def main(args: Array[String]): Unit = {
+    // #hello-world
+    val system: ActorSystem[HelloWorldMain.SayHello] =
+      ActorSystem(HelloWorldMain(), "hello")
+
+    system ! HelloWorldMain.SayHello("World")
+    system ! HelloWorldMain.SayHello("Akka")
+    // #hello-world
+
+    Thread.sleep(3000)
+    system.terminate()
+  }
+  //#hello-world-main
+}
+//#hello-world-main
