+++ samples/akka-quickstart-scala/src/test/scala/com/example/HelloWorldSpec.scala
@@ -1,0 +1,19 @@
+package com.example
+
+import akka.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
+import com.example.HelloWorld.Greet
+import com.example.HelloWorld.Greeted
+import org.scalatest.wordspec.AnyWordSpecLike
+
+class HelloWorldSpec extends ScalaTestWithActorTestKit with AnyWordSpecLike {
+
+  "A Greeter" must {
+    "reply to greeted" in {
+      val replyProbe = createTestProbe[Greeted]()
+      val underTest = spawn(HelloWorld())
+      underTest ! Greet("Santa", replyProbe.ref)
+      replyProbe.expectMessage(Greeted("Santa", underTest.ref))
+    }
+  }
+
+}
